requirements.txt
python-telegram-bot==20.5
gspread==6.0.2
apscheduler==3.10.4
pytz==2023.3